Specs for MongoDB up and running again
