+++ Specifications/Clients/MongoDB/Hosting/for_MongoDBReadModels/when_adding_read_models_from_provider.cs
@@ -1,23 +1,18 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Aksio.DependencyInversion;
-using Aksio.Types;
+using Aksio.Cratis.MongoDB;
 using Microsoft.Extensions.DependencyInjection;
 using MongoDB.Driver;
 
 namespace Aksio.Cratis.Hosting.for_MongoDBReadModels;
 
-public class when_adding_read_models_with_type_taking_provider_for_collection : Specification
+public class when_adding_read_models_from_provider : Specification
 {
     record ReadModel();
-    class TypeWithDependencies
-    {
-        public TypeWithDependencies(ProviderFor<IMongoCollection<ReadModel>> provider) { }
-    }
 
     Mock<IServiceCollection> service_collection;
-    Mock<ITypes> types;
+    Mock<ICanProvideMongoDBReadModelTypes> types;
 
     ServiceDescriptor service_descriptor;
 
@@ -26,10 +21,10 @@ public class when_adding_read_models_with_type_taking_provider_for_collection : 
         service_collection = new();
         service_collection.Setup(_ => _.Add(IsAny<ServiceDescriptor>())).Callback((ServiceDescriptor sp) => service_descriptor = sp);
         types = new();
-        types.SetupGet(_ => _.All).Returns(new[] { typeof(TypeWithDependencies) });
+        types.Setup(_ => _.Provide()).Returns(new[] { typeof(ReadModel) });
     }
 
-    void Because() => service_collection.Object.UseMongoDBReadModels(types.Object);
+    void Because() => service_collection.Object.UseMongoDBReadModels(readModelTypeProvider: types.Object);
 
     [Fact] void should_register_collection_provider() => service_descriptor.ServiceType.ShouldEqual(typeof(IMongoCollection<ReadModel>));
 }
